resource-scripts/es_ping/requirements.txt
cloudshell-automation-api
cloudshell-logging
pingparsing
cloudshell-sandbox-reporter